requirements.txt
Pillow>=9.0.1
python-resize-image>=1.1.20